fix: ingredient details get from params
