+++ src/components/index.js
@@ -2,7 +2,6 @@ export { App } from "./app/app";
 export { AppHeader } from "./app-header/app-header";
 export { BurgerConstructor } from "./burger-constructor/burger-constructor";
 export { BurgerIngredients } from "./burger-ingredients/burger-ingredients";
-export { IngredientDetails } from "./ingredient-details/ingredient-details";
 export { OrderDetails } from "./order-details/order-details";
 export { LayoutForm, LayoutMain, LayoutProfile } from "./layouts";
 export { ModalFullfilled, ModalPending, ModalRejected } from "./modal";
@@ -11,4 +10,7 @@ export {
   OnlyAuthorizedElement,
   OnlyUnauthorizedElement,
 } from "./protected-route-element/protected-route-element";
-export { IngredientDetailsModal } from "./ingredient-details";
+export {
+  IngredientDetailsModal,
+  IngredientDetailsParams,
+} from "./ingredient-details";
