+++ src/components/ingredient-details/index.js
@@ -1,1 +1,2 @@
+export { IngredientDetailsParams } from "./ingredient-details-params";
 export { IngredientDetailsModal } from "./ingredient-details-modal";
