--- src/components/ingredient-details/ingredient-details-state.jsx
@@ -1,9 +1,0 @@
-import { selectIngredient } from "../../services/ingredient";
-import { useSelector } from "react-redux";
-import { IngredientDetails } from "./ingredient-details";
-
-export const IngredientDetailsState = () => {
-  const ingredient = useSelector(selectIngredient);
-
-  return <IngredientDetails ingredient={ingredient} />;
-};
